refactor l1-ringbuf.cpp into header + implementation; split out l1-rpc.cpp
