+++ Makefile
@@ -82,7 +82,7 @@ ch-frb-test: ch-frb-test.cpp ch_frb_rpc.o
 ch-frb-test-debug: ch-frb-test.cpp ch_frb_rpc.o $(IO_OBJS)
 	$(CPP) -o $@ $^ $(CPP_CFLAGS) $(CPP_LFLAGS) -lzmq -lhdf5
 
-ringbuf: l1-ringbuf.cpp $(IO_OBJS)
+ringbuf: l1-ringbuf.cpp l1-rpc.cpp $(IO_OBJS)
 	$(CPP) $(CPP_CFLAGS) $(CPP_LFLAGS) -o $@ $^ -lzmq -lhdf5 -llz4
 
 clean:
